requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
gradio>=4.0.0
openai>=1.1.0
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
